Better organize things in their correct modules for better separation and make the validation of events work correctly for all cases
