+++ testrium/core.py
@@ -3,102 +3,21 @@ import time
 import importlib.util
 import toml
 from colorama import init, Fore, Style
-import shutil
-from .utils import Events_Manager
+from .modules.events import Events_Manager
+from .common.loaders import load_config, load_special_callbakcs, load_test_functions
+from .common.utils import print_banner
+from .common.utils import suppress_output
+import pandas as pd
 
 # Initialize colorama
 init(autoreset=True)
 
-# Loads the config for each test:
-def load_config(config_path):
-    print(f"{Fore.CYAN}Loading config from {config_path}")
-    with open(config_path, 'r') as file:
-        config = toml.load(file)
-    return config
-
-# Loads the test functions:
-def load_test_functions(dir_path):
-    test_functions = {}
-    for filename in os.listdir(dir_path):
-        if filename.startswith('_') or not filename.endswith('.py'):
-            continue
-        module_name = filename[:-3]
-        file_path = os.path.join(dir_path, filename)
-        spec = importlib.util.spec_from_file_location(module_name, file_path)
-        module = importlib.util.module_from_spec(spec)
-        spec.loader.exec_module(module)
-        for attr in dir(module):
-            if attr == "Events_Manager":
-                continue
-            if attr == "extra_test_validation":
-                continue
-            if attr == "test_results_handler":
-                continue
-            func = getattr(module, attr)
-            if callable(func) and not attr.startswith('_'):
-                test_functions[attr] = func
-    return test_functions
-
-def load_special_callbakcs (dir_path):
-    
-    """
-    This method is used to load the special functions that are:
-    - extra_test_validation
-    - test_results_handler
-    
-    This helper callbacks allows more dinamic patterns to emerge
-    """
-    
-    def get_fn (module, attr):
-        func = getattr(module, attr)
-        if callable(func) and not attr.startswith('_'):
-            return func
-    
-    special_callbacks = {
-        "extra_test_validation": None,
-        "test_results_handler": None,
-    }
-    
-    for filename in os.listdir(dir_path):
-        if filename.startswith('_') or not filename.endswith('.py'):
-            continue
-        module_name = filename[:-3]
-        file_path = os.path.join(dir_path, filename)
-        spec = importlib.util.spec_from_file_location(module_name, file_path)
-        module = importlib.util.module_from_spec(spec)
-        spec.loader.exec_module(module)
-        
-        for attr in dir(module):
-            if attr == "extra_test_validation":
-                special_callbacks[attr] = get_fn(module, attr)
-            elif attr == "test_results_handler":
-                special_callbacks[attr] = get_fn(module, attr)
-            else:
-                continue
-    
-    return special_callbacks
-        
 # Run setup of the test group
 def run_setup(setup_path):
     spec = importlib.util.spec_from_file_location("setup", setup_path)
     setup_module = importlib.util.module_from_spec(spec)
     spec.loader.exec_module(setup_module)
     setup_module.main()
-    
-# Print the nice Pass or Fail banners
-def print_banner(message, color=Fore.GREEN):
-    term_width = shutil.get_terminal_size().columns
-    message_length = len(message)
-    padding_length = (term_width - message_length) // 2
-    banner = "=" * padding_length + message + "=" * padding_length
-
-    # Ensure the banner fits exactly in the terminal width
-    if len(banner) < term_width:
-        banner += "=" * (term_width - len(banner))
-
-    print(color + banner)
-    
-from .utils import suppress_output
 
 def main():
     base_dir = os.getcwd()
@@ -260,8 +179,12 @@ def main():
             unit_name = unit["name"]
             if unit_name == "":
                 continue
-            unit_events = Events_Manager(Unit=unit_name, path=base_dir).List_Events()
+            
+            unit_events = pd.DataFrame.from_dict(Events_Manager(Unit=unit_name, path=base_dir).List_Events()).loc[:, "StepCompleted"].to_list()
+            
+            print(f"{unit_events}")
             for event in eval(unit["events"]):
+                print(f"Event: {event}")
                 if event not in unit_events:
                     print(f"event: [{event}] was not completed!")
                     events_missing.append(event)
