+++ testrium/modules/events.py
@@ -1,43 +1,11 @@
-import time
-
-def log_test_time(func):
-    def wrapper(*args, **kwargs):
-        start_time = time.time()
-        result = func(*args, **kwargs)
-        elapsed_time = time.time() - start_time
-        print(f"{func.__name__} took {elapsed_time:.2f} seconds")
-        return result
-    return wrapper
-
-def verify_condition(condition, message="Verification failed"):
-    if not condition:
-        raise AssertionError(message)
-
-import contextlib
-import os
-import sys
-
-@contextlib.contextmanager
-def suppress_output():
-    with open(os.devnull, 'w') as devnull:
-        old_stdout = sys.stdout
-        old_stderr = sys.stderr
-        sys.stdout = devnull
-        sys.stderr = devnull
-        try:
-            yield
-        finally:
-            sys.stdout = old_stdout
-            sys.stderr = old_stderr
-
-
 # TODO >>> Add a Events Manager that can save events locally
 # Events are temporary milestones that we can define to a test 
 # This allows us to log events that we expect the test to complete.
 
-from .common.sql_pool import SQLiteConnectionPool
+from ..common.sql_pool import SQLiteConnectionPool
 import os
 import pandas as pd
+import time
 
 THIS_DIR = os.getcwd()
