+++ tests/test_2/test_case.py
@@ -1,4 +1,4 @@
-from testrium.utils import log_test_time, verify_condition
+from testrium.common.utils import log_test_time, verify_condition
 
 @log_test_time
 def test_example():
